fix: analyze Wasm plugin version without instantiating plugins
