+++ crates/dprint/src/plugins/implementations/wasm/compile.rs
@@ -2,10 +2,9 @@ use anyhow::bail;
 use anyhow::Result;
 
 use super::create_identity_import_object;
-use super::functions::WasmFunctions;
+use super::create_wasm_plugin_instance;
 use super::load_instance::load_instance;
 use super::load_instance::WasmModuleCreator;
-use super::InitializedWasmPluginInstance;
 use crate::plugins::CompilationResult;
 
 /// Compiles a Wasm module.
@@ -13,7 +12,7 @@ pub fn compile(wasm_bytes: &[u8]) -> Result<CompilationResult> {
   let wasm_module_creator = WasmModuleCreator::default();
   let module = wasm_module_creator.create_from_wasm_bytes(wasm_bytes)?;
 
-  let bytes = match module.inner.serialize() {
+  let bytes = match module.inner().serialize() {
     Ok(bytes) => bytes,
     Err(err) => bail!("Error serializing wasm module: {:#}", err),
   };
@@ -22,8 +21,7 @@ pub fn compile(wasm_bytes: &[u8]) -> Result<CompilationResult> {
   let mut store = wasmer::Store::default();
   let imports = create_identity_import_object(&mut store);
   let instance = load_instance(&mut store, &module, &imports)?;
-  let wasm_functions = WasmFunctions::new(store, instance)?;
-  let mut instance = InitializedWasmPluginInstance::new(wasm_functions)?;
+  let mut instance = create_wasm_plugin_instance(store, instance)?;
 
   Ok(CompilationResult {
     bytes: bytes.into(),
