+++ crates/dprint/src/plugins/implementations/wasm/load_instance.rs
@@ -8,10 +8,20 @@ use wasmer::Instance;
 use wasmer::Module;
 use wasmer::Store;
 
+use super::instance::get_current_plugin_schema_version;
+use super::PluginSchemaVersion;
+
 #[derive(Clone)]
 pub struct WasmInstance {
   pub inner: wasmer::Instance,
   pub engine: wasmer::Engine,
+  version: PluginSchemaVersion,
+}
+
+impl WasmInstance {
+  pub fn version(&self) -> PluginSchemaVersion {
+    self.version
+  }
 }
 
 /// Loads a compiled wasm module from the specified bytes.
@@ -21,6 +31,7 @@ pub fn load_instance(store: &mut Store, module: &WasmModule, import_object: &Imp
     Ok(instance) => Ok(WasmInstance {
       inner: instance,
       engine: module.engine.clone(),
+      version: module.version,
     }),
     Err(err) => bail!("Error instantiating module: {:#}", err),
   }
@@ -28,8 +39,23 @@ pub fn load_instance(store: &mut Store, module: &WasmModule, import_object: &Imp
 
 #[derive(Clone)]
 pub struct WasmModule {
-  pub inner: wasmer::Module,
+  inner: wasmer::Module,
   engine: wasmer::Engine,
+  version: PluginSchemaVersion,
+}
+
+impl WasmModule {
+  pub fn new(module: wasmer::Module, engine: wasmer::Engine) -> Result<Self> {
+    Ok(Self {
+      version: get_current_plugin_schema_version(&module)?,
+      inner: module,
+      engine,
+    })
+  }
+
+  pub fn inner(&self) -> &wasmer::Module {
+    &self.inner
+  }
 }
 
 // https://github.com/wasmerio/wasmer/pull/3378#issuecomment-1327679422
@@ -49,20 +75,15 @@ impl Default for WasmModuleCreator {
 impl WasmModuleCreator {
   pub fn create_from_wasm_bytes(&self, wasm_bytes: &[u8]) -> Result<WasmModule> {
     let engine_ref = EngineRef::new(&self.engine);
-    Ok(WasmModule {
-      inner: Module::new(&engine_ref, wasm_bytes)?,
-      engine: self.engine.clone(),
-    })
+    let module = Module::new(&engine_ref, wasm_bytes)?;
+    WasmModule::new(module, self.engine.clone())
   }
 
   pub fn create_from_serialized(&self, compiled_module_bytes: &[u8]) -> Result<WasmModule> {
     unsafe {
       let engine_ref = EngineRef::new(&self.engine);
       match Module::deserialize(&engine_ref, compiled_module_bytes) {
-        Ok(module) => Ok(WasmModule {
-          inner: module,
-          engine: self.engine.clone(),
-        }),
+        Ok(module) => WasmModule::new(module, self.engine.clone()),
         Err(err) => bail!("Error deserializing compiled wasm module: {:#}", err),
       }
     }
