+++ crates/dprint/src/plugins/implementations/wasm/mod.rs
@@ -1,13 +1,13 @@
 mod compile;
-mod functions;
 mod import_object;
+mod instance;
 mod load_instance;
 mod plugin;
 mod setup_wasm_plugin;
 
 pub use compile::*;
-use functions::*;
 pub use import_object::*;
+use instance::*;
 pub use load_instance::WasmModuleCreator;
 use load_instance::*;
 pub use plugin::*;
